Initial code provided by David
This code has one or two minor comments. The README.md is the read me file that goes along with this project. The only thing missing is the node_modules folder.  The code can be run by using node app.js. This code will potentially change depending on our needs.